requirements.txt
requests==2.31.0
python-dotenv==1.0.0
# Testing dependencies
pytest==7.4.3
pytest-mock==3.12.0
